move system tests to a separate module
